Sort out scripts into a hierarchy structure
The scripts folder is very messy, so I have created subfolders to put
scripts in based on what they do. Hopefully the structure is good, we
can refine it later, and here is just a starting point.

It looks like there is also some redundant scripts that we will need to
clean out.

I won't add all of the asset move changes in this message because there
are a lot, but look at the commit files changed for more information.