Move prototype level sprites into environment folder
It seems like it's better placed here. However we'll need a better
environment hierarchy for when we start doing the UI redesign and making
the real levels.

An idea is to have folders for each environment object, so that it is
easy to look up and we can pair the animations and spites together.