Remove Animations folder, refactor assets
I moved the sprites/animations from the Animations folder to where they
should go in the heirarchy of sprites. This is because currently it
seems a bit silly to have sprites that have animations in it's own
subfolder, and this makes it harder to find specific sprites. Also some
sprites had animations, but weren't in the Animations/ subfolder. Hence,
I have removed this folder and moved the sprite assets to where they
should go.