Cleanup prototype (#55)
After completing the prototype, the project was a bit messy. I have
gone through and cleaned up some of the files, delete duplicates and
unused files, as well as moving assets around for better cohesion,
searchability etc.

I have not touched most of the script, we will still need to tidy these
up.