Change bug collectable script to be bound to collectables
From learning that Pep has a lot of collision events (due to him, as the
playable character, being in contact with the ground and walls and other
environmental assets), it is better to have the collision event for
Pep<->Collectable bug attached to the bug. I have refactored the current
implementation so that on collision, the score is incremented and the
bug destroys itself, instead of Pep destroying the collided bug.