Implement centralised game manager (#71)
Dylan and I went through and unified all of the manager as they were
very separated and could be combined. This makes it easier to serialise
and generate save states (and also load states in). Also this will help
with extensibility of the manager.

The HealthManager, ScoreManager and ItemManager have been combined into
one GameManager. This now holds all of the game's state. Load and save
functionality has been implemented in GameManager also which is
responsible for storing and retrieving game state from file. This
GameManager has a traditional OOP singleton style as opposed to a
MonoBehaviour Unity one.

The Score Management has been slightly rewritten, with the collision
detection associated with the collectables instead of Pep.

Minor fixes were also introduces to some of the scenes from the
prototype, and prefabs relating to the old manager style have been
removed.