Enhance GameManager; Implement factory scene generator/switcher (#76)
I have implemented more OOP appropriate level factory generator and
switcher, which utilises both factory and template patterns. Hooks can
be added so that upon scene changes, we can execute code to reset game
state, for example.

There is also a Scene class hierarchy that has been introduced. The
superclass Scene is subclassed (one per scene asset), and each child
will have it's own hooks for start up and tear down so when the scene is
switched in or out, we can execute code that will set the game state to
something stable. These classes is in a sub namespace POCC.Scenes, and
I've also refactored the other Scene related classes/enums in there too.

An adapter SceneSwitcherAdapter was implemented so that the events
within the scenes can use the scene switching mechanism.

The prototype scenes (the 4 of them, I didn't include the swingtest
scene) have been changed to use this new scene switching implementation.